+++ example_project/config/settings.py
@@ -43,8 +43,8 @@ INSTALLED_APPS = (
     'rest_framework.authtoken',
     'social_django',
     'rest_social_auth',
-    'sslserver',  # some social providers require https
     'knox',  # For django-rest-knox
+    'django_extensions',  # some social providers require https
 
     'users',
 )