+++ tests/test_session.py
@@ -7,7 +7,7 @@ from django.test import modify_settings
 from django.test.utils import override_settings
 from rest_framework.test import APITestCase
 from rest_framework.permissions import DjangoModelPermissionsOrAnonReadOnly
-from httpretty import HTTPretty
+import responses
 from social_core.utils import parse_qs
 
 from .base import BaseFacebookAPITestCase, BaseTwitterApiTestCase
@@ -43,6 +43,7 @@ class TestSocialAuth2(APITestCase, BaseFacebookAPITestCase):
 
     @modify_settings(**session_modify_settings)
     def _check_login_social_session(self, url, data):
+        self.setup_api_mocks()
         resp = self.client.post(url, data)
         self.assertEqual(resp.status_code, 200)
         self.assertEqual(resp.data['email'], self.email)
@@ -75,31 +76,34 @@ class TestSocialAuth2(APITestCase, BaseFacebookAPITestCase):
         self.assertEqual(resp.status_code, 404)
 
     def test_login_social_http_origin(self):
+        self.setup_api_mocks()
         resp = self.client.post(
             reverse('login_social_session'),
             data={'provider': 'facebook', 'code': '3D52VoM1uiw94a1ETnGvYlCw'},
             HTTP_ORIGIN="http://frontend.com")
         self.assertEqual(resp.status_code, 200)
-        url_params = dict(parse_qsl(urlparse(HTTPretty.latest_requests[0].path).query))
+        url_params = dict(parse_qsl(urlparse(responses.calls[0].request.path_url).query))
         self.assertEqual(url_params['redirect_uri'], "http://frontend.com/")
 
     @override_settings(REST_SOCIAL_OAUTH_ABSOLUTE_REDIRECT_URI='http://myproject.com/')
     def test_login_absolute_redirect(self):
+        self.setup_api_mocks()
         resp = self.client.post(
             reverse('login_social_session'),
             data={'provider': 'facebook', 'code': '3D52VoM1uiw94a1ETnGvYlCw'})
         self.assertEqual(resp.status_code, 200)
-        url_params = dict(parse_qsl(urlparse(HTTPretty.latest_requests[0].path).query))
+        url_params = dict(parse_qsl(urlparse(responses.calls[0].request.path_url).query))
         self.assertEqual('http://myproject.com/', url_params['redirect_uri'])
 
     @override_settings(REST_SOCIAL_OAUTH_ABSOLUTE_REDIRECT_URI='http://myproject.com/')
     def test_login_manual_redirect(self):
+        self.setup_api_mocks()
         resp = self.client.post(
             reverse('login_social_session'),
             data={'provider': 'facebook', 'code': '3D52VoM1uiw94a1ETnGvYlCw',
                   'redirect_uri': 'http://manualdomain.com/'})
         self.assertEqual(resp.status_code, 200)
-        url_params = dict(parse_qsl(urlparse(HTTPretty.latest_requests[0].path).query))
+        url_params = dict(parse_qsl(urlparse(responses.calls[0].request.path_url).query))
         self.assertEqual('http://manualdomain.com/', url_params['redirect_uri'])
 
     @mock.patch('rest_framework.views.APIView.permission_classes')